More work on deb release scripts, and a necessary change in GUI to avoid compile problems in Ubuntu

+++ HOPSAN++/HopsanGUI/GUIObjects/GUISystem.h
@@ -139,7 +139,7 @@ protected:
 private:
     void commonConstructorCode();
 
-    size_t mNumberOfLogSamples;
+    int mNumberOfLogSamples;
 
     QString mLoadType;
     CoreSystemAccess *mpCoreSystemAccess;
